+++ lib/view_component/errors.rb
@@ -43,17 +43,6 @@ module ViewComponent
     end
   end
 
-  class DuplicateContentError < StandardError
-    MESSAGE =
-      "It looks like a block was provided after calling `with_content` on COMPONENT, " \
-      "which means that ViewComponent doesn't know which content to use.\n\n" \
-      "To fix this issue, use either `with_content` or a block."
-
-    def initialize(klass_name)
-      super(MESSAGE.gsub("COMPONENT", klass_name.to_s))
-    end
-  end
-
   class MissingCollectionArgumentError < StandardError
     MESSAGE =
       "The initializer for COMPONENT doesn't accept the parameter `PARAMETER`, " \